% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-assets.R
\name{geo_assets}
\alias{geo_assets}
\title{Obtain plotly's geo assets as an htmlDependency object}
\usage{
geo_assets()
}
\description{
This function is used by plotly to make it possible to render
maps without an internet connection.
}
\examples{

geo_assets()

}
\author{
Carson Sievert
}
